model layer done

+++ model/CourseRepository.js
@@ -10,7 +10,7 @@ class CourseRepository extends Course{
     try{
       const result = await db.query("INSERT INTO Course(title, description, instructorName) VALUES($1, $2, $3)",
       [this.title, this.description, this.instructorName]);
-      return res.rows[0];
+      return result.rows[0];
     }catch(err){
       return err.stack;
     }
@@ -20,7 +20,7 @@ class CourseRepository extends Course{
     try{
       const result = await db.query("SELECT * FROM Course WHERE idCourse=$1",
       [this.id]);
-      return res.rows[0];
+      return result.rows[0];
     }catch(err){
       return err.stack;
     }
@@ -29,10 +29,18 @@ class CourseRepository extends Course{
   async updateColumn(column, newValue){
     let query = "";
 
+    const data = await this.selectAllData();
+
+    if(!data){
+      return "404";
+    }
+
     if(column === "title"){
       query = "UPDATE Course SET title=$1 WHERE idCourse=$2";
     }else if(column === "description"){
       query = "UPDATE Course SET description=$1 WHERE idCourse=$2";
+    }else{
+      return "Column does not exist or you are not alowed to modify it";
     }
 
     try{
@@ -44,6 +52,12 @@ class CourseRepository extends Course{
   }
 
   async deleteCourse(){
+    const data = this.selectAllData();
+
+    if(!data){
+      return "404";
+    }
+
     try{
       const result = await db.query("DELETE FROM Course WHERE idCourse=$1",
       [this.id]);
@@ -57,7 +71,8 @@ class CourseRepository extends Course{
     try{
       const result = await db.query("SELECT * FROM Course WHERE title=$1",
         [this.title]);
-      return res.rows[0];
+
+      return result.rows[0];
     }catch(err){
       return err.stack;
     }
