+++ model/StudentRepository.js
@@ -19,6 +19,9 @@ class StudentRepository extends Student{
   async selectAllData(){
     try{
       const result = await db.query("SELECT * FROM Student WHERE idStudent=$1", [this.id]);
+
+      this.name = result.rows[0].name;
+
       return result.rows[0];
     }catch(err){
       return err.stack;
@@ -38,7 +41,7 @@ class StudentRepository extends Student{
     try{
       const result = await db.query("SELECT currentCourses FROM Student WHERE idStudent=$1",
       [this.id]);
-      return result.rows[0];
+      return result.rows[0].currentcourses;
     }catch(err){
       return err.stack;
     }
@@ -47,17 +50,25 @@ class StudentRepository extends Student{
   async updateColumn(column, newValue){
     let query = "";
 
+    const id = await this.selectIDAndPass();
+
+    if(!id){
+      return "404";
+    }
+
     if(column === "name"){
       query = "UPDATE Student SET name=$1 WHERE idStudent=$2";
     }else if(column === "bio"){
       query = "UPDATE Student SET biography=$1 WHERE idStudent=$2";
     }else if(column === "password"){
       query = "UPDATE Student SET password=$1 WHERE idStudent=$2";
+    }else{
+      return "Column does not exists or you are not allowed to modify it";
     }
 
     try{
       const result = await db.query(query, [newValue, this.id]);
-      return "Updated: "+res.rows[0];
+      return "Updated: "+result.rows[0];
     }catch(err){
       return err.stack;
     }
@@ -65,6 +76,12 @@ class StudentRepository extends Student{
 
   async updateCoursesArray(changeOnlyEnd, newValue){
     let query = "";
+
+    const id = await this.selectIDAndPass();
+
+    if(!id){
+      return "404";
+    }
 
     if(changeOnlyEnd){
       query = "UPDATE Student SET currentCourses=array_append(currentCourses, $1) WHERE idStudent=$2";
@@ -81,9 +98,15 @@ class StudentRepository extends Student{
   }
 
   async deleteStudent(){
+    const id = await this.selectIDAndPass();
+
+    if(!id){
+      return "404";
+    }
+
     try{
       const result = await db.query("DELETE FROM Student WHERE idStudent=$1", [this.id]);
-      return res.rows[0];
+      return result.rows[0];
     }catch(err){
       return err.stack;
     }
